refactor post-build script to ensure files that already have .user.js do not have additional .user.js extensions added to them.

+++ post-build.sh
@@ -3,8 +3,8 @@
 # rename all .js files in dist/ to .user.js to that tampermonkey can recognize
 # them as user scripts and prompt for installation
 for file in dist/*.js; do
-    if [ -f "$file" ]; then
-        # echo "Renaming $file to ${file%.js}.user.js"
+    if [ -f "$file" ] && [ "${file%.user.js}" = "$file" ]; then
+        # echo "renaming $file to ${file%.js}.user.js"
         mv "$file" "${file%.js}.user.js"
     fi
 done
